requirements.txt
python-dotenv
python-telegram-bot==21.11.1
aiohttp
aiosqlite
APScheduler
